Moved LoadingSpinner into separate component folder
